Update README.md
The code a DYNAMIC API SMART CONTRACT REACTIONARY ALGORITHM FOR USDSHARES (USDS) NON-VOLATILE STABLECOIN/STABLETOKEN provided is a Solidity smart contract that implements a reactionary algorithm for the USDSHARES stabletoken. 

The contract USDSReactionaryAlgorithm contains several public variables to store the prices of different mainstream cryptocurrency assets, including USDSPrice, USDSHARESPrice, CRYTOSHARESPrice, BITCOINPrice, ETHEREUMPrice, BNBPrice, MATICPrice, and TRONPrice. These variables are used to track the prices of the respective digital currencies.

In the constructor function, the initial prices of all the tokens are set to 1. This is the starting point for the algorithm.

The reactToSupplyDemand function takes two parameters: supplyChange and demandChange. If either of these parameters is greater than or equal to 10, it means there has been a significant change in supply or demand. In such a case, the algorithm increases the USDSHARESPrice by 5% (multiplies it by 105 and divides by 100). This increase in price reflects the reaction to the change in supply or demand.

The reactToMarketPriceChange function takes six parameters: crytoSharesPriceChange, bitcoinPriceChange, ethereumPriceChange, bnbPriceChange, maticPriceChange, and tronPriceChange. If any of these parameters is less than 0, it means the price of the corresponding token has decreased. In this scenario, the algorithm increases the USDSHARESPrice by $100,000. This increase in price is a reaction to the market price change.

Overall, this USDS API algorithmic solidity smart contract implements a reactionary algorithm for USDSHARES(USDS) stabletoken, adjusting its price based on changes in supply, demand, and market prices of other mainstream cryptocurrency assets.